refactor(ping_pong/game): ping_pong game refactoring (#21) fix(security): Security fix in clang-format action feat(README): Removing unnecessary links

+++ ping_pong/game/src/gameobj/playercontrol.cpp
@@ -1,7 +1,9 @@
 #include "gameobj/player.hpp"
 #include <ncurses.h>
 
-int Player::bot(float posY) const { return posY - this->height / 2; }
+int Player::bot(float posY) const {
+    return posY - static_cast<float>(this->height) / 2;
+}
 
 int Player::keyboard() const {
     int result = this->pos;
